use bigIncrements insted of increments

+++ database/migrations/2018_12_17_093252_create_reset_password_tables.php
@@ -14,7 +14,7 @@ class CreateResetPasswordTables extends Migration
     public function up()
     {
         Schema::create('api_password_reset_codes', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('email')->index()->nullable();
             $table->string('mobile')->index()->nullable();
             $table->string('code');
@@ -22,8 +22,8 @@ class CreateResetPasswordTables extends Migration
         });
 
         Schema::create('api_password_reset_tokens', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('user_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
             $table->string('token');
             $table->timestamp('updated_at');
