+++ database/migrations/2018_4_10_171249_create_service_tokens_table.php
@@ -14,10 +14,10 @@ class CreateServiceTokensTable extends Migration
     public function up()
     {
         Schema::create('service_tokens', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('name');
             $table->string('token', 500);
-            $table->unsignedInteger('user_id');
+            $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
             $table->timestamps();
         });
